Configure SSL bundle with reload on update
This commit configures the given key and certificate with the server and
allows to reload automatically the certificate when it's updated on
disk.

To showcase this behavior, the app now runs on port 8443 with HTTPs. We
can see its current certificate by invoking the following script:

$ scripts/get-certs.sh

We can then update the cert using "scripts/update-certs.sh", wait for
the app to pick it up and get the certs again.